% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantlyzer.R
\name{all_class_together_scaled}
\alias{all_class_together_scaled}
\title{My function description}
\usage{
all_class_together_scaled(data, pos_X, pos_y, method, scale_method)
}
\description{
My function description
}
