% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantlyzer.R
\name{print_result}
\alias{print_result}
\title{My function description}
\usage{
print_result(
  KNN_result,
  naive_result,
  svm_result,
  rf_result,
  xgb_result,
  adab_result,
  keras_result
)

print_result(
  KNN_result,
  naive_result,
  svm_result,
  rf_result,
  xgb_result,
  adab_result,
  keras_result
)
}
\description{
My function description

My function description
}
